# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather & Auth API",
        "description": "Documentation for the Weather & Auth API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8004"
        },
        {
            "url": "http://ms-weather.sytes.net"
        }
    ],
    "paths": {
...
        "/api/v1/weather-module/weather/by-city": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Get weather data by city",
                "operationId": "3033db15d54dafd73a006d8a6d4389fc",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City name to retrieve weather data",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language for API responses (en or es)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "es"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Weather data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Weather data retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "location": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "London"
                                                        },
                                                        "region": {
                                                            "type": "string",
                                                            "example": "City of London, Greater London"
                                                        },
                                                        "country": {
                                                            "type": "string",
                                                            "example": "United Kingdom"
                                                        },
                                                        "localtime": {
                                                            "type": "string",
                                                            "example": "2024-03-15 15:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "current": {
                                                    "properties": {
                                                        "temp_c": {
                                                            "type": "number",
                                                            "example": "10"
                                                        },
                                                        "temp_f": {
                                                            "type": "number",
                                                            "example": "50"
                                                        },
                                                        "condition": {
                                                            "properties": {
                                                                "text": {
                                                                    "type": "string",
                                                                    "example": "Partly cloudy"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "wind_kph": {
                                                            "type": "number",
                                                            "example": "15"
                                                        },
                                                        "humidity": {
                                                            "type": "number",
                                                            "example": "60"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid data"
                                        },
                                        "data": {
                                            "properties": {
                                                "city": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invalid city name"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error retrieving weather data"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Error message from the service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}